Ejercicios

+++ src/Ej05.py
@@ -1,6 +1,7 @@
-#Escribe un programa que pida el importe sin IVA de un artículo y el tipo de IVA a aplicar y calcule e imprima por pantalla el precio final del artículo.
+#Escribe un programa que pida el importe sin IVA de un artículo y el tipo de IVA a aplicar y calcule e imprima por pantalla el precio 
+# final del artículo.
 
-num1 = int(input("Dime el coste del artículo\n"))
+num1 = float(input("Dime el coste del artículo\n"))
 
 num2 = int(input("Dime el porcentaje de IVA del artículo\n"))
